switched to -g flag for @mouse_drag_status
I previously set this as a server setting with -s and it worked fine.
Until I whilst working on something else ran tmux customize-mode
tmux instantly crashed with "server exited unexpectedly".
It turned out this was caused by putting this user option with -s
tmux didn't show any error or warning when setting an -s user variable,
and it could be read fine with -s  :(